requirements.txt
anyio==3.4.0
argon2-cffi==21.1.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
ecdsa==0.17.0
fastapi==0.70.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.12.0
hyperframe==6.0.1
idna==3.3
passlib==1.7.4
priority==2.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-jose==3.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
typing-extensions==4.0.1
wsproto==1.0.0
